axis_label move class to separate file
