+++ srcts/display_scatter/display_scatter.ts
@@ -12,6 +12,7 @@ import {
 import { Timeline } from "./timeline";
 import { centerColumns, getColMeans } from "./utils";
 import { SelectionHelper } from "./selection_helper";
+import { AxisLabel } from "./axis_label";
 import { OrbitControls } from "three/examples/jsm/controls/OrbitControls";
 import { brushIcon, orbitIcon, panIcon, resetIcon, selectIcon } from "./icons";
 import "./style.css";
@@ -871,60 +872,5 @@ export abstract class DisplayScatter {
       this.orbitControls.enabled = true;
       this.selectionHelper.disable();
     }
-  }
-}
-
-class AxisLabel {
-  private div: HTMLDivElement;
-  private canvas: HTMLCanvasElement;
-  private text: string;
-  private position: THREE.Vector3;
-  private dim: Dim;
-  private dpr: number;
-
-  constructor(
-    text: string,
-    pos: number[],
-    container: HTMLElement,
-    canvas: HTMLCanvasElement,
-    camera: Camera,
-    dim: Dim,
-    dpr: number,
-  ) {
-    this.div = document.createElement("div");
-    this.div.innerHTML = text;
-    this.div.className = "axisLabel";
-
-    this.canvas = canvas;
-    this.text = text;
-    this.position = new THREE.Vector3();
-    this.dim = dim;
-    this.dpr = dpr;
-
-    container.appendChild(this.div);
-    this.updatePosition(pos, camera);
-  }
-
-  public updatePosition(pos: number[], camera: Camera) {
-    if (this.dim == 3) {
-      this.position.set(pos[0], pos[1], pos[2]);
-    } else {
-      this.position.set(pos[0], 0, pos[1]);
-    }
-    var coords2d = this.get2DCoords(camera);
-
-    this.div.style.left = coords2d.x + "px";
-    this.div.style.top = coords2d.y + "px";
-  }
-
-  public setDpr(dpr: number) {
-    this.dpr = dpr;
-  }
-
-  private get2DCoords(camera: Camera) {
-    var vector = this.position.project(camera);
-    vector.x = (vector.x + 1) * this.canvas.width / (2 * this.dpr);
-    vector.y = -(vector.y - 1) * this.canvas.height / (2 * this.dpr);
-    return vector;
   }
 }
